# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Personas y Aulas",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/direcciones": {
            "get": {
                "tags": [
                    "Direcciones"
                ],
                "summary": "Listar direcciones",
                "operationId": "5804759279a63322458aea4bbde01e2b",
                "responses": {
                    "200": {
                        "description": "Lista de direcciones"
                    }
                }
            },
            "post": {
                "tags": [
                    "Direcciones"
                ],
                "summary": "Crear direccion",
                "operationId": "a8d160b54e268d0a7cc954733f788b15",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "calle",
                                    "numero",
                                    "id_ciudad"
                                ],
                                "properties": {
                                    "calle": {
                                        "type": "string",
                                        "example": "Calle 1"
                                    },
                                    "numero": {
                                        "type": "integer",
                                        "example": 101
                                    },
                                    "id_ciudad": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Direccion creada"
                    }
                }
            }
        },
        "/api/direcciones/{id}": {
            "put": {
                "tags": [
                    "Direcciones"
                ],
                "summary": "Actualizar direccion",
                "operationId": "ec17a1a1a38f1afc31806f94c89a0b44",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la direccion",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "calle",
                                    "numero",
                                    "id_ciudad"
                                ],
                                "properties": {
                                    "calle": {
                                        "type": "string",
                                        "example": "Calle 1"
                                    },
                                    "numero": {
                                        "type": "integer",
                                        "example": 101
                                    },
                                    "id_ciudad": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Direccion actualizada"
                    },
                    "404": {
                        "description": "Direccion no encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Direcciones"
                ],
                "summary": "Eliminar direccion",
                "operationId": "ba12ad8725b2ea572c0a7409772ee75b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la direccion",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Direccion eliminada"
                    },
                    "404": {
                        "description": "Direccion no encontrada"
                    }
                }
            }
        },
...
        "/api/grados": {
            "get": {
                "tags": [
                    "Grados"
                ],
                "summary": "Listar grados",
                "operationId": "d8d5bea7d7ccf255046bfafccd582c62",
                "responses": {
                    "200": {
                        "description": "Lista de grados"
                    }
                }
            },
            "post": {
                "tags": [
                    "Grados"
                ],
                "summary": "Crear grado",
                "operationId": "c89c65f572f53c52615db173936122cc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "id_profesor"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Grado 1"
                                    },
                                    "id_profesor": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Grado creada"
                    }
                }
            }
        },
        "/api/grados/{id}": {
            "put": {
                "tags": [
                    "Grados"
                ],
                "summary": "Actualizar grado",
                "operationId": "9afd4729054e8d9ba7dea405c450b682",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la grado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "id_profesor"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Grado 1"
                                    },
                                    "id_profesor": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Grado actualizada"
                    },
                    "404": {
                        "description": "Grado no encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Grados"
                ],
                "summary": "Eliminar grado",
                "operationId": "8b1f1f150ec0732ca99d483d40476d73",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la grado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Grado eliminada"
                    },
                    "404": {
                        "description": "Grado no encontrada"
                    }
                }
            }
        },
...
    },
...
}